refactoring of thoughts_json_parser + add class parser
