+++ requirements.txt
@@ -1,1 +1,3 @@
 langchain==0.0.202
+setuptools==67.8.0
+pydantic==1.10.9
