+++ yid_langchain_extensions/output_parser/thoughts_json_parser.py
@@ -2,7 +2,7 @@ import json
 from typing import List, Dict, Any
 
 from langchain import PromptTemplate
-from langchain.schema import AgentAction, BaseOutputParser
+from langchain.schema import BaseOutputParser
 from pydantic import BaseModel
 
 FORMAT_PROMPT = """FORMAT:
@@ -52,19 +52,3 @@ class ThoughtsJSONParser(BaseOutputParser):
         format_instructions = PromptTemplate.from_template(FORMAT_PROMPT, template_format="jinja2").format_prompt(
             thoughts=self.format_thoughts()).to_string()
         return format_instructions
-
-
-class ActionParser(ThoughtsJSONParser):
-    extra_thoughts: List[Thought] = []
-    action_thoughts: List[Thought] = [
-        Thought(name="action", description="The action to take. Must be one of [{tool_names}]"),
-        Thought(name="action_input", description="The input to the action")
-    ]
-
-    @property
-    def thoughts(self) -> List[Thought]:
-        return self.action_thoughts + self.extra_thoughts
-
-    def parse(self, text: str) -> AgentAction:
-        response = super().parse(text)
-        return AgentAction(response["action"], response["action_input"], text)
